requirements.txt
streamlit==1.37.0
deepface==0.0.88
tensorflow==2.15.1
tf-keras==2.15.1
chromadb==0.5.3
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.2
opencv-python==4.9.0.80
